+++ src/wai/common/json/configuration/property/_AnyOfProperty.py
@@ -1,14 +1,14 @@
 from typing import List, Optional
 
-from ...schema import one_of
+from ...schema import any_of
 from ._OfProperty import OfProperty
 from ._Property import Property
 from ._DummyInstance import DummyInstance
 
 
-class OneOfProperty(OfProperty):
+class AnyOfProperty(OfProperty):
     """
-    Property which validates JSON that matches exactly one of
+    Property which validates JSON that matches at least one of
     a number of schema.
     """
     def __init__(self,
@@ -18,23 +18,15 @@ class OneOfProperty(OfProperty):
         super().__init__(
             name,
             *sub_properties,
-            schema_function=one_of,
+            schema_function=any_of,
             optional=optional
         )
 
     def choose_current_property(self, keys: List[Optional[DummyInstance]]) -> int:
-        # Start with an invalid index
-        index = -1
-
         # Search the list
         for i, key in enumerate(keys):
+            # Return the first valid index found
             if key is not None:
-                if index >= 0:
-                    raise AttributeError(f"Value matched more than one sub-property")
+                return i
 
-                index = i
-
-        if index == -1:
-            raise AttributeError(f"Value didn't match any sub-properties")
-
-        return index
+        raise AttributeError(f"Value didn't match any sub-properties")