+++ src/wai/common/json/configuration/property/_BoolProperty.py
@@ -1,9 +1,8 @@
 from ...schema import BOOL_SCHEMA
-from .._OptionallyPresent import OptionallyPresent
-from ._Property import Property
+from ._RawProperty import RawProperty
 
 
-class BoolProperty(Property):
+class BoolProperty(RawProperty):
     """
     Configuration property which validates a boolean value.
     """
@@ -16,9 +15,3 @@ class BoolProperty(Property):
             BOOL_SCHEMA,
             optional=optional
         )
-
-    def get_as_raw_json(self, instance) -> OptionallyPresent[bool]:
-        return super().get_as_raw_json(instance)
-
-    def set_from_raw_json(self, instance, value: OptionallyPresent[bool]):
-        super().set_from_raw_json(instance, value)