+++ src/wai/common/json/configuration/property/_ConstantProperty.py
@@ -1,10 +1,9 @@
 from ..._typing import RawJSONPrimitive
-from .._OptionallyPresent import OptionallyPresent
-from ._Property import Property
 from ...schema import constant
+from ._RawProperty import RawProperty
 
 
-class ConstantProperty(Property[RawJSONPrimitive]):
+class ConstantProperty(RawProperty):
     """
     Configuration property which validates a constant primitive value.
     """
@@ -18,9 +17,3 @@ class ConstantProperty(Property[RawJSONPrimitive]):
             constant(value),
             optional=optional
         )
-
-    def get_as_raw_json(self, instance) -> OptionallyPresent[RawJSONPrimitive]:
-        return super().get_as_raw_json(instance)
-
-    def set_from_raw_json(self, instance, value: OptionallyPresent[RawJSONPrimitive]):
-        super().set_from_raw_json(instance, value)