+++ src/wai/common/json/configuration/property/__init__.py
@@ -1,6 +1,7 @@
 """
 Package for the different types of properties a JSON configuration can have.
 """
+from ._AnyOfProperty import AnyOfProperty
 from ._ArrayProperty import ArrayProperty
 from ._BoolProperty import BoolProperty
 from ._ConstantProperty import ConstantProperty
@@ -9,4 +10,5 @@ from ._NumberProperty import NumberProperty
 from ._OneOfProperty import OneOfProperty
 from ._Property import Property
 from ._ProxyProperty import ProxyProperty
+from ._RawProperty import RawProperty
 from ._StringProperty import StringProperty