+++ CHANGES.rst
@@ -49,3 +49,4 @@ Changelog
 - Added exception package to meta package for processing exceptions. Currently only
   contains ExceptionChainer, which captures exceptions as a context-manager and then
   provides methods for processing them.
+- Added utility interface JSONValidatedBiserialisable.